+++ tests/test_brier.py
@@ -25,3 +25,15 @@ class TestBrier:
         )
         brier = calculators.Brier(true_alternative_index=1,)
         assert brier.calculate(prediction) == decimal.Decimal("0.735")
+
+    def test_order_matters(self):
+        prediction = predictions.Prediction(
+            probabilities=[
+                decimal.Decimal(25),
+                decimal.Decimal(25),
+                decimal.Decimal(30),
+                decimal.Decimal(20),
+            ],
+        )
+        ordered_categorical = calculators.OrderedCategorical(true_alternative_index=1)
+        assert ordered_categorical.calculate(prediction) == decimal.Decimal("0.2350")